requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
greenlet==3.0.3
idna==3.6
pdfkit==1.0.0
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
